Small changes.
[refix] I'm stupid. Tax stuff was right in the first place.
[modified] Just so this commit isn't wasted just to cover up my stupidity, I'm throwing in modifications made to other classes that don't need the Mission Manager.

git-svn-id: http://svn2.assembla.com/svn/swgemu/core3devel/MMOCoreORB@493 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/managers/player/PlayerManagerImplementation.cpp
@@ -784,7 +784,7 @@ void PlayerManagerImplementation::moveItem(Player* sender, Player* receiver, Tan
 
 void PlayerManagerImplementation::doBankTip(Player* sender, Player* receiver, uint32 tipAmount, bool updateTipTo) {
 	//Pre: sender wlocked
-	uint32 tax = tipAmount * .05;
+	float tax = tipAmount * .05;
 	if (!sender->verifyBankCredits(tipAmount + tax)) {
 		sender->sendSystemMessage("You lack the required funds to do that. (Bank Tip.)");
 		return;
